+++ R/adrr.R
@@ -1,7 +1,7 @@
 #' Calculate average daily risk range (ADRR)
 #'
 #' @description
-#' The function adrr produces ADRR values in a tibble object.
+#' The function `adrr` produces ADRR values in a tibble object.
 #'
 #' @usage
 #' adrr(data)
@@ -15,7 +15,7 @@
 #'
 #' @details
 #' A tibble object with 1 row for each subject, a column for subject id and
-#' a column for ADRR values is returned. NA glucose values are
+#' a column for ADRR values is returned. `NA` glucose values are
 #' omitted from the calculation of the ADRR values.
 #'
 #' ADRR is the average sum of HBGI corresponding to the highest glucose